+++ Yank/schema_tools/__init__.py
@@ -1,0 +1,5 @@
+"""
+A set of tools for validating Cerberus-based schemas
+The modules of this sub-package are NOT public, but helper utilities.
+The public facing
+"""